backend/requirements.txt

fastapi==0.115.12
uvicorn==0.34.2

pandas==2.2.3
numpy==2.0.2 
langchain==0.3.25
langchain-core==0.3.60
langchain-google-genai==2.1.4

python-dotenv==1.1.0 
python-multipart==0.0.20
